Updated Reactivate page to show the navbar
